20/04/2023 - 8th commit
included analysis of male and female butterfly forewing length on june mean separately.

deleted lm relating to jun_mean and year since it is not a causation relationship but a correlation one.

used cor.test to test correlation of jun_mean and year and tested its assumptions in the data import script